refactor(ReferentielCard, ExpandableAction): reorganize imports and update layout structure for better responsiveness
